extracted KumatoProvider
